Refactor latent inference network components and tests
