+++ bayesflow/networks/__init__.py
@@ -20,7 +20,8 @@ from .inference import ConsistencyModel, StableConsistencyModel
 from .inference import CouplingFlow
 from .inference import DiffusionModel
 from .inference import FlowMatching
-from .inference import LatentDiffusionModel, Encoder, Decoder
+from .inference import LatentInferenceNetwork, Encoder, Decoder
+from .inference import LatentDiffusionModel
 from .inference import PointNetwork, ScoringRuleNetwork
 
 # Summary networks
