+++ bayesflow/networks/inference/__init__.py
@@ -7,7 +7,8 @@ from .consistency import ConsistencyModel, StableConsistencyModel
 from .coupling import CouplingFlow
 from .diffusion import DiffusionModel
 from .flow_matching import FlowMatching
-from .latent_diffusion import LatentDiffusionModel, Encoder, Decoder
+from .latent import LatentInferenceNetwork, Encoder, Decoder
+from .latent_diffusion import LatentDiffusionModel
 from .scoring import ScoringRuleNetwork, PointNetwork
 from .scoring import (
     ScoringRule,
