+++ bayesflow/networks/inference/latent/__init__.py
@@ -1,0 +1,7 @@
+from .latent_inference_network import LatentInferenceNetwork
+from .encoder import Encoder
+from .decoder import Decoder
+
+from bayesflow.utils._docs import _add_imports_to_all
+
+_add_imports_to_all()
