+++ bayesflow/networks/inference/latent_diffusion/__init__.py
@@ -1,6 +1,5 @@
 from .latent_diffusion_model import LatentDiffusionModel
-from .encoder import Encoder
-from .decoder import Decoder
+from ..latent import Encoder, Decoder
 
 from bayesflow.utils._docs import _add_imports_to_all
