+++ tests/test_networks/test_latent_inference_network/test_encoder.py
